feat: add test path in remove old symlink

+++ home/.chezmoiscripts/run_after_executable_symlinks.ps1
@@ -6,7 +6,9 @@ sudo {
       [string]$source
     )
     try {
-      Remove-Item -Path $target -Force -Recurse -ErrorAction Stop
+      if (Test-Path $target) {
+        Remove-Item -Path $target -Force -Recurse -ErrorAction Stop
+      }
       New-Item -ItemType SymbolicLink -Path $target -Target $source -ErrorAction Stop | Out-Null
     }
     catch {
